:recycle: Update auth
