+++ src/components/Promotion/organisms/index.tsx
@@ -1,9 +1,9 @@
-import useLogin from 'hook/useLogin'
 import { useRouter } from 'next/navigation'
 import { gauthUrl } from 'utils/libs/gauthUrl'
 import GauthLoginButton from '../atom/GauthLoginButton'
 import Header from '../atom/Header'
 import { Wrapper } from './style'
+import { useLogin } from 'utils/apis/auth'
 
 const Promotion = () => {
   const router = useRouter()
