+++ src/utils/apis/auth.ts
@@ -1,6 +1,10 @@
 import { GetServerSidePropsContext } from 'next'
 import { apiClient } from 'utils/libs/apiClient'
 import { setToken } from 'utils/libs/setToken'
+import { useRouter, useSearchParams } from 'next/navigation'
+import { useEffect } from 'react'
+import { TokensType } from 'types/apis/TokenManager'
+import { toast } from 'react-toastify'
 export const tokenReissue = async (
   refreshToken: string,
   ctx: GetServerSidePropsContext | null,
@@ -21,4 +25,28 @@ export const tokenReissue = async (
     setToken(newAcessToken, refreshToken, ctx)
     return { newAcessToken }
   } catch (e: any) {}
+}
+
+export const useLogin = () => {
+  const router = useRouter()
+
+  const params = useSearchParams()
+  const gauthCode = params.get('code')
+
+  const fetch = async (data: { code: string }) => {
+    try {
+      const { data: tokens } = await apiClient.post<TokensType>('auth', data)
+      setToken(tokens.accessToken, tokens.refreshToken, null)
+      router.push('/home')
+      toast.success('로그인 성공')
+    } catch (e: any) {
+      router.push('/')
+      toast.error('로그인 실패')
+    }
+  }
+
+  useEffect(() => {
+    if (!gauthCode) return
+    fetch({ code: gauthCode })
+  }, [gauthCode])
 }
